+++ src/constants.js
@@ -336,6 +336,7 @@ export const CONFIGURATIONS_ENTRIES_API = `${CONFIGURATIONS_API}/entries`;
 export const CONSORTIA_API = 'consortia';
 export const CONSORTIA_TENANTS_API = 'tenants';
 export const CONSORTIA_USER_TENANTS_API = 'user-tenants';
+export const GROUPS_API = 'groups';
 export const PROFILE_PIC_API = 'users/profile-picture';
 export const USERS_API = 'users';
 export const PATRON_PREREGISTRATIONS_API = 'staging-users';